Include veteran information in HCA authenticated payload
