--- src/applications/hca/api/mockapi.js
@@ -1,63 +1,0 @@
-/* eslint-disable camelcase */
-const commonResponses = require('../../../platform/testing/local-dev-mock-api/common');
-
-module.exports = {
-  ...commonResponses,
-  //   'GET path/to/endpoint': { foo: 'bar' },
-  'GET /v0/user': {
-    data: {
-      attributes: {
-        profile: {
-          sign_in: {
-            service_name: 'idme',
-          },
-          email: 'fake@fake.com',
-          loa: { current: 3 },
-          first_name: 'Jane',
-          middle_name: '',
-          last_name: 'Doe',
-          gender: 'F',
-          birth_date: '1985-01-01',
-          verified: true,
-        },
-        veteran_status: {
-          status: 'OK',
-          is_veteran: true,
-          served_in_military: true,
-        },
-        in_progress_forms: [],
-        prefills_available: ['21-526EZ'],
-        services: [
-          'facilities',
-          'hca',
-          'edu-benefits',
-          'evss-claims',
-          'form526',
-          'user-profile',
-          'health-records',
-          'rx',
-          'messaging',
-        ],
-        va_profile: {
-          status: 'OK',
-          birth_date: '19511118',
-          family_name: 'Hunter',
-          gender: 'M',
-          given_names: ['Julio', 'E'],
-          active_status: 'active',
-          facilities: [
-            {
-              facility_id: '983',
-              is_cerner: false,
-            },
-            {
-              facility_id: '984',
-              is_cerner: false,
-            },
-          ],
-        },
-      },
-    },
-    meta: { errors: null },
-  },
-};
