+++ src/applications/hca/config/chapters/veteranInformation/americanIndian.js
@@ -1,7 +1,8 @@
 import fullSchemaHca from 'vets-json-schema/dist/10-10EZ-schema.json';
 import { AmericanIndianDescription } from '../../../components/FormDescriptions';
 import { ShortFormAlert } from '../../../components/FormAlerts';
-import { emptyObjectSchema, NotHighDisability } from '../../../helpers';
+import { NotHighDisability } from '../../../utils/helpers';
+import { emptyObjectSchema } from '../../../definitions';
 
 const { sigiIsAmericanIndian } = fullSchemaHca.properties;
