+++ src/applications/hca/config/chapters/veteranInformation/birthInformation.js
@@ -3,8 +3,8 @@ import constants from 'vets-json-schema/dist/constants.json';
 
 import AuthenticatedShortFormAlert from '../../../components/FormAlerts/AuthenticatedShortFormAlert';
 import { BirthInfoDescription } from '../../../components/FormDescriptions';
-
-import { emptyObjectSchema, NotHighDisability } from '../../../helpers';
+import { NotHighDisability } from '../../../utils/helpers';
+import { emptyObjectSchema } from '../../../definitions';
 
 const { cityOfBirth } = fullSchemaHca.properties;
 const { states50AndDC } = constants;
