+++ src/applications/hca/config/chapters/veteranInformation/birthSex.js
@@ -4,7 +4,8 @@ import { genderLabels } from 'platform/static-data/labels';
 
 import { ShortFormAlert } from '../../../components/FormAlerts';
 import CustomReviewField from '../../../components/FormReview/CustomReviewField';
-import { emptyObjectSchema, NotHighDisability } from '../../../helpers';
+import { NotHighDisability } from '../../../utils/helpers';
+import { emptyObjectSchema } from '../../../definitions';
 
 const { gender } = fullSchemaHca.properties;
