+++ src/applications/hca/config/chapters/veteranInformation/maidenNameInformation.js
@@ -2,7 +2,8 @@ import fullSchemaHca from 'vets-json-schema/dist/10-10EZ-schema.json';
 import set from 'platform/utilities/data/set';
 
 import { ShortFormAlert } from '../../../components/FormAlerts';
-import { emptyObjectSchema, NotHighDisability } from '../../../helpers';
+import { NotHighDisability } from '../../../utils/helpers';
+import { emptyObjectSchema } from '../../../definitions';
 
 const { mothersMaidenName } = fullSchemaHca.properties;
