+++ src/applications/hca/config/chapters/veteranInformation/veteranAddress.js
@@ -8,7 +8,8 @@ import {
 
 import { MailingAddressDescription } from '../../../components/FormDescriptions';
 import { ShortFormAlert } from '../../../components/FormAlerts';
-import { emptyObjectSchema, NotHighDisability } from '../../../helpers';
+import { NotHighDisability } from '../../../utils/helpers';
+import { emptyObjectSchema } from '../../../definitions';
 
 export default {
   uiSchema: {
