+++ src/applications/hca/definitions/dependent.js
@@ -5,7 +5,7 @@ import fullNameUI from 'platform/forms/definitions/fullName';
 import currentOrPastDateUI from 'platform/forms-system/src/js/definitions/currentOrPastDate';
 import ssnUI from 'platform/forms-system/src/js/definitions/ssn';
 import currencyUI from 'platform/forms-system/src/js/definitions/currency';
-import { validateDependentDate } from '../validation';
+import { validateDependentDate } from '../utils/validation';
 
 const incomeFields = ['grossIncome', 'netIncome', 'otherIncome'];
