+++ src/applications/hca/definitions/index.js
@@ -1,0 +1,5 @@
+// define empty schema for view fields
+export const emptyObjectSchema = {
+  type: 'object',
+  properties: {},
+};
