+++ src/applications/hca/tests/helpers.unit.spec.js
@@ -3,14 +3,9 @@ import { expect } from 'chai';
 import {
   didEnrollmentStatusChange,
   expensesLessThanIncome,
-  getCSTOffset,
-  getOffsetTime,
-  getAdjustedTime,
-  isAfterCentralTimeDate,
-  isBeforeCentralTimeDate,
   transformAttachments,
   prefillTransformer,
-} from '../helpers';
+} from '../utils/helpers';
 
 describe('HCA helpers', () => {
   describe('expensesLessThanIncome', () => {
@@ -93,40 +88,6 @@ describe('HCA helpers', () => {
         .be.false;
       expect(expensesLessThanIncome('deductibleFuneralExpenses')(formData)).to
         .be.true;
-    });
-  });
-  describe('getCSTOffset', () => {
-    it('should return -300 if is daylight savings time', () => {
-      expect(getCSTOffset(true)).to.equal(-300);
-    });
-    it('should return -360 if is not daylight savings time', () => {
-      expect(getCSTOffset(false)).to.equal(-360);
-    });
-  });
-  describe('getOffsetTime', () => {
-    it('should convert an offset number of minutes into milliseconds', () => {
-      expect(getOffsetTime(1)).to.equal(60000);
-    });
-  });
-  describe('getAdjustedTime', () => {
-    it('should determine utc time', () => {
-      expect(getAdjustedTime(1, 1)).to.equal(2);
-    });
-  });
-  describe('isAfterCentralTimeDate', () => {
-    it('should return true if the discharge date is after the Central Time reference date', () => {
-      expect(isAfterCentralTimeDate('9999-12-24')).to.be.true;
-    });
-    it('should return false if the discharge date is not after the Central Time reference date', () => {
-      expect(isAfterCentralTimeDate('2000-12-12')).to.be.false;
-    });
-  });
-  describe('isBeforeCentralTimeDate', () => {
-    it('should return true if the discharge date is after the Central Time reference date', () => {
-      expect(isBeforeCentralTimeDate('9999-12-24')).to.be.false;
-    });
-    it('should return false if the discharge date is not after the Central Time reference date', () => {
-      expect(isBeforeCentralTimeDate('2000-12-12')).to.be.true;
     });
   });
   describe('transformAttachments', () => {
