+++ src/applications/hca/tests/validation.unit.spec.js
@@ -7,7 +7,7 @@ import {
   validateDependentDate,
   // validateMarriageDate,
   validateCurrency,
-} from '../validation';
+} from '../utils/validation';
 
 describe('hca validation', () => {
   describe('validateServiceDates', () => {
