+++ src/applications/hca/utils/actions.js
@@ -2,7 +2,7 @@ import appendQuery from 'append-query';
 import { apiRequest } from 'platform/utilities/api';
 import environment from 'platform/utilities/environment';
 import recordEvent from 'platform/monitoring/record-event';
-import { getData, isServerError, isClientError } from './util';
+import { getData, isServerError, isClientError } from '.';
 import { HCA_ENROLLMENT_STATUSES } from './constants';
 import {
   dismissedHCANotificationDate,
