+++ src/applications/hca/utils/constants.js
@@ -48,3 +48,31 @@ export const IS_SHORT_FORM_ENABLED = 'isShortFormFeatureToggleEnabled';
 
 // declare states without medical care serivces
 export const statesWithoutService = ['AA', 'AE', 'AP', 'FM', 'MH', 'PW'];
+
+// declare labels for discharge type select box
+export const dischargeTypeLabels = {
+  honorable: 'Honorable',
+  general: 'General',
+  other: 'Other Than Honorable',
+  'bad-conduct': 'Bad Conduct',
+  dishonorable: 'Dishonorable',
+  undesirable: 'Undesirable',
+};
+
+// declare labels for last service branch select box
+export const lastServiceBranchLabels = {
+  'air force': 'Air Force',
+  army: 'Army',
+  'coast guard': 'Coast Guard',
+  'marine corps': 'Marine Corps',
+  'merchant seaman': 'Merchant Seaman',
+  navy: 'Navy',
+  noaa: 'Noaa',
+  'space force': 'Space Force',
+  usphs: 'USPHS',
+  'f.commonwealth': 'Filipino Commonwealth Army',
+  'f.guerilla': 'Filipino Guerilla Forces',
+  'f.scouts new': 'Filipino New Scout',
+  'f.scouts old': 'Filipino Old Scout',
+  other: 'Other',
+};
